backend/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
ciso8601==2.1.3
click==7.1.2
cryptography==2.9.2
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.0.6
django-common-helpers==0.9.2
django-cors-headers==3.3.0
django-extensions==2.2.9
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.9
joblib==0.15.1
natsort==7.0.1
nltk==3.5
oauthlib==3.1.0
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
sqlparse==0.3.1
tqdm==4.46.0
tweepy==3.8.0
tweet-preprocessor==0.6.0
urllib3==1.25.9
vaderSentiment==3.3.1
wcwidth==0.2.4
whitenoise==5.1.0
wit==6.0.0
